\name{RunNimbleParallel}
\alias{RunNimbleParallel}
\title{
  NIMBLE wrapper with parallel processing and automated sampling until convergence
}
\description{
  Implements Bayesian model fitting using the NIMBLE R packaged in parallelized clusters across multiple processors. Additionally, convergence and sampling metrics (i.e., Rhat and n_effective) are checked periodically, and sampling resumes until these metrics meet specified minimum levels. If sampling resumes, saved samples are dumped to the hard drive to avoid overloading RAM with successive rounds of resuming sampling.
  
  ***Note: This function will probably not run in Windows because parallel processing packages do not operate easily in a Windows environment.
}
\usage{
  RunNimbleParallel(model, inits, data, constants, parameters,
           nc = 2, ni = 2000, nb = 0.5, nt = 10, mod.nam = "mod",
           max.samples.saved = 10000, rtrn.model = F,
           Rht.required = 1.1, neff.required = 100)
}
\arguments{
  \item{model}{Object containing NIMBLE model code generated by the function nimble::nimbleCode.}
  \item{inits}{Function that generates initial values for stochastic nodes as required to run a NIMBLE model.}
  \item{data}{List containing all data objects required to fit specified NIMBLE model.}
  \item{constants}{List containing all constants required to fit specified NIMBLE model.}
  \item{nc}{Number of MCMC chains to run. Note available cores must equal or exceed 'nc'. Default = 2.}
  \item{ni}{Number of iterations to run MCMC chains before checking convergence and sampling metrics. Default = 10000. If model does not meet specified levels for R_hat and n_effective, sampling will continue for another 'ni' steps before checking metrics again.}
  \item{nb}{Proportion of MCMC chains to discard as burn-in.}
  \item{nt}{Base level of thinning to implement within parallelized cluster. Default = 10. Setting this argument will reduce load on RAM for large models.}
  \item{mod.nam}{Character string used as a file label for saved model outputs.}
  \item{max.samples.saved}{Integer that sets minimum level of samples to save per MCMC chain in final model object. Additional thinning of posterior samples (i.e., after samples are recovered from processing clusters) is implemented as necessary to limit the size of the model object to (roughly) this level. Default = 10000.}
  \item{rtrn.model}{Logical indicating whether to return model output to R environment. Default = F, i.e., model output is only saved to disk.}
  \item{Rht.required}{Maximum numeric value for Rhat for any parameter required to end sampling.}
  \item{neff.required}{Minimum value for n_effective for any parameter required to end sampling.}
}
\value{If assigned to object and rtrn.model = T, model output as returned by mcmcOutput::mcmcOutput is returned. Else, model output is saved to file.}
\author{
  Quresh S. Latif, Bird Conservancy of the Rockies
}

\keyword{nimble wrapper}
\keyword{mcmcOutput}
\keyword{parallel processing}
